Split read out too
